requirements.txt
Flask==2.0.1
PyMySQL==0.9.3
cryptography==2.8
pycryptodomex==3.10.1
flask-cors==3.0.8
requests==2.25.1
ffmpeg-python